Reduce bigint executor work
The `BytePolyProgram` only needs to be run during preflight.
