+++ risc0/circuit/rv32im/src/prove/witgen/preflight.rs
@@ -24,7 +24,7 @@ use risc0_zkp::core::digest::DIGEST_WORDS;
 
 use crate::{
     execute::{
-        bigint::{BigIntBytes, BigIntState},
+        bigint::BigIntBytes,
         node_idx,
         pager::{page_idx, PageState, PagedMemory},
         platform::*,
@@ -37,7 +37,10 @@ use crate::{
     zirgen::circuit::ExtVal,
 };
 
-use super::{node_addr_to_idx, node_idx_to_addr, paged_map::PagedMap, poseidon2::Checksum};
+use super::{
+    bigint::BigIntState, node_addr_to_idx, node_idx_to_addr, paged_map::PagedMap,
+    poseidon2::Checksum,
+};
 
 #[derive(Clone, Debug, Default)]
 pub(crate) enum Back {
@@ -430,6 +433,18 @@ impl<'a> Preflight<'a> {
         // tracing::trace!("add_cycle_special(cur_state: {cur_state:?}, next_state: {next_state:?}, major: {major}, minor: {minor})");
         self.add_cycle(next_state, pc, major, minor, paging_idx, back);
     }
+
+    pub(crate) fn on_bigint_cycle(&mut self, cur_state: CycleState, bigint: &BigIntState) {
+        self.add_witness(&bigint.bytes);
+        self.add_cycle_special(
+            cur_state,
+            bigint.next_state,
+            self.pc.0,
+            0, //paging_idx
+            Back::BigInt(bigint.clone()),
+        );
+        self.phys_cycles += 1;
+    }
 }
 
 impl Risc0Context for Preflight<'_> {
@@ -631,16 +646,8 @@ impl Risc0Context for Preflight<'_> {
         self.phys_cycles += 1;
     }
 
-    fn on_bigint_cycle(&mut self, cur_state: CycleState, bigint: &BigIntState) {
-        self.add_witness(&bigint.bytes);
-        self.add_cycle_special(
-            cur_state,
-            bigint.next_state,
-            self.pc.0,
-            0, //paging_idx
-            Back::BigInt(bigint.clone()),
-        );
-        self.phys_cycles += 1;
+    fn ecall_bigint(&mut self) -> Result<()> {
+        super::bigint::ecall_preflight(self)
     }
 }
